Break out additional component for list based canvas navigation
