+++ __tests__/src/components/WindowSideBarCanvasPanel.test.js
@@ -4,8 +4,7 @@ import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
 import manifesto from 'manifesto.js';
 import { WindowSideBarCanvasPanel } from '../../../src/components/WindowSideBarCanvasPanel';
-import SidebarIndexThumbnail from '../../../src/containers/SidebarIndexThumbnail';
-import SidebarIndexCompact from '../../../src/containers/SidebarIndexCompact';
+import SidebarIndexList from '../../../src/containers/SidebarIndexList';
 import CompanionWindow from '../../../src/containers/CompanionWindow';
 import manifestJson from '../../fixtures/version-2/019.json';
 
@@ -32,20 +31,10 @@ function createWrapper(props) {
 }
 
 describe('WindowSideBarCanvasPanel', () => {
-  let setCanvas;
-
-  beforeEach(() => {
-    setCanvas = jest.fn();
-  });
-
-  it('renders all needed elements for the thumbnail view', () => {
+  it('renders SidebarIndexList', () => {
     const wrapper = createWrapper();
     expect(wrapper.find(CompanionWindow).props().title).toBe('canvasIndex');
-    expect(wrapper.find(List).length).toBe(1);
-    expect(wrapper.find(ListItem).length).toBe(3);
-    expect(wrapper.find(ListItem).first().props().component).toEqual('li');
-    expect(wrapper.find(ListItem).at(1).props().selected).toBe(true);
-    expect(wrapper.find(List).find(SidebarIndexThumbnail).length).toBe(3);
+    expect(wrapper.find(SidebarIndexList).length).toBe(1);
   });
 
   describe('handleVariantChange', () => {
@@ -54,53 +43,6 @@ describe('WindowSideBarCanvasPanel', () => {
       const wrapper = createWrapper({ updateVariant });
       wrapper.instance().handleVariantChange({ target: { value: 'compact' } });
       expect(updateVariant).toHaveBeenCalledWith('compact');
-    });
-  });
-
-  it('renders all needed elements for the compact view', () => {
-    const wrapper = createWrapper({ variant: 'compact' });
-    expect(wrapper.find(CompanionWindow).props().title).toBe('canvasIndex');
-    expect(wrapper.find(List).length).toBe(1);
-    expect(wrapper.find(ListItem).length).toBe(3);
-    expect(wrapper.find(ListItem).first().props().component).toEqual('li');
-    expect(wrapper.find(List).find(SidebarIndexCompact).length).toBe(3);
-  });
-
-  it('should call the onClick handler of a list item', () => {
-    const wrapper = createWrapper({ setCanvas });
-    wrapper.find(ListItem).at(1).simulate('click');
-    expect(setCanvas).toHaveBeenCalledTimes(1);
-  });
-
-  describe('getIdAndLabelOfCanvases', () => {
-    it('should return id and label of each canvas in manifest', () => {
-      const canvases = manifesto
-        .create(manifestJson)
-        .getSequences()[0]
-        .getCanvases();
-      const wrapper = createWrapper({ canvases });
-      const received = wrapper.instance().getIdAndLabelOfCanvases(canvases);
-      const expected = [
-        {
-          id: 'http://iiif.io/api/presentation/2.0/example/fixtures/canvas/24/c1.json',
-          label: 'Test 19 Canvas: 1',
-        },
-        {
-          id: 'https://purl.stanford.edu/fr426cg9537/iiif/canvas/fr426cg9537_1',
-          label: 'Image 1',
-        },
-        {
-          id: 'https://purl.stanford.edu/rz176rt6531/iiif/canvas/rz176rt6531_1',
-          label: 'Image 2',
-        },
-      ];
-      expect(received).toEqual(expected);
-    });
-
-    it('should return empty array if canvas if empty', () => {
-      const wrapper = createWrapper({ canvases: [] });
-      const received = wrapper.instance().getIdAndLabelOfCanvases([]);
-      expect(received).toEqual([]);
     });
   });
 });
