+++ src/components/WindowSideBarCanvasPanel.js
@@ -1,17 +1,12 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import Typography from '@material-ui/core/Typography';
-import List from '@material-ui/core/List';
-import ListItem from '@material-ui/core/ListItem';
 import MenuItem from '@material-ui/core/MenuItem';
 import FormControl from '@material-ui/core/FormControl';
 import RootRef from '@material-ui/core/RootRef';
 import Select from '@material-ui/core/Select';
-import { ScrollTo } from './ScrollTo';
-import ManifestoCanvas from '../lib/ManifestoCanvas';
 import CompanionWindow from '../containers/CompanionWindow';
-import SidebarIndexCompact from '../containers/SidebarIndexCompact';
-import SidebarIndexThumbnail from '../containers/SidebarIndexThumbnail';
+import SidebarIndexList from '../containers/SidebarIndexList';
 
 /**
  * a panel showing the canvases for a given manifest
@@ -30,16 +25,6 @@ export class WindowSideBarCanvasPanel extends Component {
   }
 
   /** @private */
-  getIdAndLabelOfCanvases() {
-    const { canvases } = this.props;
-
-    return canvases.map((canvas, index) => ({
-      id: canvas.id,
-      label: new ManifestoCanvas(canvas).getLabel(),
-    }));
-  }
-
-  /** @private */
   handleVariantChange(event) {
     const { updateVariant } = this.props;
 
@@ -52,11 +37,8 @@ export class WindowSideBarCanvasPanel extends Component {
    */
   render() {
     const {
-      canvases,
       classes,
       id,
-      selectedCanvases,
-      setCanvas,
       t,
       toggleDraggingEnabled,
       variant,
@@ -64,7 +46,6 @@ export class WindowSideBarCanvasPanel extends Component {
     } = this.props;
 
     const { variantSelectionOpened } = this.state;
-    const canvasesIdAndLabel = this.getIdAndLabelOfCanvases(canvases);
     return (
       <RootRef rootRef={this.containerRef}>
         <CompanionWindow
@@ -103,35 +84,7 @@ export class WindowSideBarCanvasPanel extends Component {
             </FormControl>
           )}
         >
-          <List>
-            {
-              canvasesIdAndLabel.map((canvas, canvasIndex) => {
-                const onClick = () => { setCanvas(windowId, canvas.id); }; // eslint-disable-line require-jsdoc, max-len
-
-                return (
-                  <ScrollTo
-                    containerRef={this.containerRef}
-                    key={`${canvas.id}-${variant}`}
-                    offsetTop={96} // offset for the height of the form above
-                    scrollTo={!!selectedCanvases.find(c => c.id === canvas.id)}
-                  >
-                    <ListItem
-                      key={canvas.id}
-                      className={classes.listItem}
-                      alignItems="flex-start"
-                      onClick={onClick}
-                      button
-                      component="li"
-                      selected={!!selectedCanvases.find(c => c.id === canvas.id)}
-                    >
-                      {variant === 'compact' && <SidebarIndexCompact canvas={canvas} />}
-                      {variant === 'thumbnail' && <SidebarIndexThumbnail canvas={canvas} otherCanvas={canvases[canvasIndex]} />}
-                    </ListItem>
-                  </ScrollTo>
-                );
-              })
-            }
-          </List>
+          <SidebarIndexList id={id} containerRef={this.containerRef} windowId={windowId} />
         </CompanionWindow>
       </RootRef>
     );
@@ -139,11 +92,8 @@ export class WindowSideBarCanvasPanel extends Component {
 }
 
 WindowSideBarCanvasPanel.propTypes = {
-  canvases: PropTypes.array.isRequired, // eslint-disable-line react/forbid-prop-types
   classes: PropTypes.objectOf(PropTypes.string).isRequired,
   id: PropTypes.string.isRequired,
-  selectedCanvases: PropTypes.arrayOf(PropTypes.shape({ id: PropTypes.string })),
-  setCanvas: PropTypes.func.isRequired,
   t: PropTypes.func.isRequired,
   toggleDraggingEnabled: PropTypes.func.isRequired,
   updateVariant: PropTypes.func.isRequired,
@@ -152,6 +102,5 @@ WindowSideBarCanvasPanel.propTypes = {
 };
 
 WindowSideBarCanvasPanel.defaultProps = {
-  selectedCanvases: [],
   variant: 'thumbnail',
 };
